docs: add introduction to repository
This was made to be an inital seed to develop a ledger that is able to contain all aspects of SWE rationalizing within one coherent framework. This work is by no means final, and is a WIP for life. Will I continue? Who knows. I have omitted the mention of the details of the inner frameworks that I've alluded to in the introduction. This is for a few reasons: 1) I didn't know how to structure them within the frameworks, and I didn't have any ideas for that. 2) I felt that even if included, they would make the introduction too lengthy, and therefore chose to not even consider thinking about how to structure them. 3) I'm not mature or fully knowledgeable of how licensing works, So I had a few concerns on the nature of my frameworks, and the sources I've been inspired by for these frameworks. Now I do realize that this work is not influential enough to be under any legal risk even if I did break some rules, but this is a learning opporunity for me to experience thinkingabout these topics, and taking them into consideration. Starting small will help when things actually become serious.

This was my full rationale for it, if I missed answering any questions relevant to this tasks, it is simply due to my ignorance of said questions, and the scenarios that call for said questions. But I'm learning. :)